mac_requirements.txt
blinker==1.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Mail==0.9.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
six==1.15.0
Werkzeug==1.0.1
